backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
python-dotenv==1.0.1
google-genai==1.32.0
pydantic==2.9.2
PyPDF2==3.0.1
